% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aggregate_mts}
\alias{aggregate_mts}
\title{Aggregate Multivariate Time Series for K-L divergence}
\usage{
aggregate_mts(x)
}
\arguments{
\item{x}{NumericMatrix multivariate time series}
}
\value{
NumericVector
}
\description{
This functions aggregates multivariate times series into univariate time series.
See details.
}
\details{
To eliminate local spikes, compute usual distance.
\deqn{\sum_{i \neq j}^p \lvert x_{ti} - x_{tj}}
where p is the number of variables, and t is the index of time.
This enables to explain the correlation between the series.
}
\references{
Cho, J., Tariq, S., Lee, S., Kim, Y. G., & Woo, S. (2019). \emph{Contextual Anomaly Detection by Correlated Probability Distributions using Kullback-Leibler Divergence}. Workshop on Mining and Learning From Time Series. \url{http://doi.org/10.1145/nnnnnnn.nnnnnnn}
}
