% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_cusum}
\alias{compute_cusum}
\title{Window CUSUM}
\usage{
compute_cusum(x, win, jump, norm, display_progress = FALSE)
}
\arguments{
\item{x}{NumericMatrix multivariate time series, which is forecasting error}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{norm}{int p-norm}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
NumericVector
}
\description{
This function computes the sum of p-norm in each window.
}
\details{
This function will be used in both \code{\link{detect_cusum}} and \code{\link{train_cusum}}
Use error data as an input.
}
\seealso{
\code{\link{detect_cusum}}
  \code{\link{train_cusum}}
}
