% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_euc}
\alias{compute_euc}
\title{Squared l2 Distance between two windows}
\usage{
compute_euc(x, y)
}
\arguments{
\item{x}{NumericMatrix. first window.}

\item{y}{NumericMatrix. second window.}
}
\value{
double. squared l2 distance matrix form.
}
\description{
This function gives distance matrix between two windows in multivariate time series
}
\details{
Compute
\deqn{d = x_{ij} - y_{kl}}
element-wise. After that,
\deqn{\frac{\sqrt{\sum{d}}}{wp}}
where w is the wize of window, and p is the number of variables.
}
\references{
Yun, J.-H., Hwang, Y., Lee, W., Ahn, H.-K., & Kim, S.-K. (2018). \emph{Statistical Similarity of Critical Infrastructure Network Traffic Based on Nearest Neighbor Distances} (Vol. 11050, pp. 1–23). Presented at the Research in Attacks, Intrusions, and Defenses, Cham: Springer International Publishing. \url{http://doi.org/10.1007/978-3-030-00470-5_27}
}
