% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_kl}
\alias{compute_kl}
\title{Kullback-Leibler divergence estimation between two densities}
\usage{
compute_kl(f1, f2)
}
\arguments{
\item{f1}{NumericMatrix density estimated by \code{\link{est_density}}. previous pdf.}

\item{f2}{NumericMatrix density estimated by \code{\link{est_density}}. current pdf.}
}
\value{
double
}
\description{
This function computes Kullback-Leibler divergence from f2 to f1.
When sliding windows, f1 is the previous pdf and f2 is the current pdf.
}
\details{
Let \eqn{\mathcal{X}} be the support of f1. Then K-L divergence from f2 to f1 is defined by
\deqn{E_{X_1} \log \frac{f_1 (x)}{f_2 (x)}}
Thus, given probability mass p and q, we can compute this K-L divergence by
\deqn{\sum_{\mathcal{X}} p(x) \log \frac{p (x)}{q (x)}}
To estimate this value, first use \code{\link{est_density}} or \code{\link{density_cpp}}, and estimate density of each window.
}
\references{
Cho, J., Tariq, S., Lee, S., Kim, Y. G., & Woo, S. (2019). \emph{Contextual Anomaly Detection by Correlated Probability Distributions using Kullback-Leibler Divergence}. Workshop on Mining and Learning From Time Series. \url{http://doi.org/10.1145/nnnnnnn.nnnnnnn}
}
\seealso{
\code{\link{est_density}}
    \code{\link{density_cpp}}
}
