% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_norm}
\alias{compute_norm}
\title{p-norm of each observation}
\usage{
compute_norm(x, norm, display_progress = FALSE)
}
\arguments{
\item{x}{NumericMatrix multivariate time series, which is forecasting error}

\item{norm}{int p-norm}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
NumericVector
}
\description{
This function computes p-norm w.r.t. the variables.
}
\seealso{
\code{\link{detect_norm}}
  \code{\link{train_norm}}
}
