% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{density_cpp}
\alias{density_cpp}
\title{Gaussian Kernel Density Estimation in C++}
\usage{
density_cpp(x)
}
\arguments{
\item{x}{NumericVector data for estimation. x of \link[stats]{density.default}.}
}
\value{
NumericMatrix of 2 columns.
First column is the n coordinates of the points where the density is estimated (x).
Second column is the estimated density values (y).
}
\description{
This function calls \link[stats]{density.default} in Rcpp.
Since it calls an R function to Rcpp, it might be slower even than R function itself.
This one is just for the sake of defining other functions in Rcpp syntax more easily.
}
\details{
Note that \link[stats]{density.default} has various arguments and options.
This function, howerver, only computes gaussian kernel.
It tries to detect windows derived from other Normal distribution.
}
\references{
Cho, J., Tariq, S., Lee, S., Kim, Y. G., & Woo, S. (2019). \emph{Contextual Anomaly Detection by Correlated Probability Distributions using Kullback-Leibler Divergence}. Workshop on Mining and Learning From Time Series. \url{http://doi.org/10.1145/nnnnnnn.nnnnnnn}
}
\seealso{
\link[stats]{density.default}
   \code{\link{est_density}}
}
