% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_cusum}
\alias{detect_cusum}
\title{CUSUM}
\usage{
detect_cusum(x, win, jump, norm, threshold, display_progress = FALSE)
}
\arguments{
\item{x}{NumericMatrix multivariate time series, which is forecasting error}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{norm}{int p-norm}

\item{threshold}{double threshold for anomaly}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
LogicalVector
}
\description{
This function detects anomaly for each window based on CUSUM.
}
\details{
If the p-norm is larger than given threshold, then the observation is anomaly.
}
