% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_mse}
\alias{detect_mse}
\title{Static Threshold based on MSE}
\usage{
detect_mse(x, win, jump, threshold)
}
\arguments{
\item{x}{NumericMatrix multivariate time series, which is forecasting error}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{threshold}{double threshold for anomaly}
}
\value{
LogicalVector
}
\description{
This function detects anomaly for each window based on window MSE.
}
\details{
If the MSE is larger than given threshold, then the window is anomaly.
}
\references{
Pavel Filonov, Andrey Lavrentyev, and Artem Vorontsov. 2016. \emph{Multivariate industrial time series with cyber-attack simulation: Fault detection using an lstm-based predictive data model}. arXiv preprint \url{arXiv:1612.06676} (2016).
}
