% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_nnd}
\alias{detect_nnd}
\title{Anomaly detection using NND}
\usage{
detect_nnd(data, win, thr)
}
\arguments{
\item{data}{NumericMatrix multivariate data set.}

\item{win}{int window size for sliding window.}

\item{thr}{double threshold that will be compared to nnd vector.}
}
\value{
LogicalVector,
If NND is (strictly) larger than threshold then TRUE.
Otherwise, FALSE
}
\description{
This function detects anomaly based on NND.
}
\details{
Given n x p data, slide a window.
Compute NND for each pair of moving window.
For threshold, users can use tail value of \code{\link{euc_pdf}}.
}
\references{
Filonov, P., Kitashov, F., & Lavrentyev, A. (2017). \emph{RNN-based Early Cyber-Attack Detection for the Tennessee Eastman Process}. CoRR.

Yun, J.-H., Hwang, Y., Lee, W., Ahn, H.-K., & Kim, S.-K. (2018). \emph{Statistical Similarity of Critical Infrastructure Network Traffic Based on Nearest Neighbor Distances} (Vol. 11050, pp. 1–23). Presented at the Research in Attacks, Intrusions, and Defenses, Cham: Springer International Publishing. \url{http://doi.org/10.1007/978-3-030-00470-5_27}
}
