% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nnd.R
\name{detect_nndvec}
\alias{detect_nndvec}
\title{Anomaly detection after conducting NND}
\usage{
detect_nndvec(nnd, win, thr, anom = c(TRUE, FALSE))
}
\arguments{
\item{nnd}{NumericVector result of \code{\link{nns_cpp}}}

\item{win}{int window size for sliding window}

\item{thr}{threshold for anomaly detection, in each window}

\item{anom}{vector represents anomaly and normal. By default, TRUE and FALSE.}
}
\value{
LogicalVector,
If NND is (strictly) larger than threshold then TRUE.
Otherwise, FALSE
}
\description{
This function detects anomaly based on NND, given \code{\link{nns_cpp}}.
}
\details{
Given n x p data, slide a window.
Compute NND for each pair of moving window.
For threshold, users can use tail value of \code{\link{euc_pdf}}.
}
\references{
Filonov, P., Kitashov, F., & Lavrentyev, A. (2017). \emph{RNN-based Early Cyber-Attack Detection for the Tennessee Eastman Process}. CoRR.

Yun, J.-H., Hwang, Y., Lee, W., Ahn, H.-K., & Kim, S.-K. (2018). \emph{Statistical Similarity of Critical Infrastructure Network Traffic Based on Nearest Neighbor Distances} (Vol. 11050, pp. 1–23). Presented at the Research in Attacks, Intrusions, and Defenses, Cham: Springer International Publishing. \url{http://doi.org/10.1007/978-3-030-00470-5_27}
}
