% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_static}
\alias{detect_static}
\title{Simple Static Threshold}
\usage{
detect_static(x, win, jump, threshold, display_progress = FALSE)
}
\arguments{
\item{x}{NumericMatrix multivariate time series, which is forecasting error}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{threshold}{double threshold for anomaly}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
LogicalVector
}
\description{
This function detects anomaly for each window using static threshold.
}
\details{
If at least one observation in the window is larger than the threshold, the entire window is anomaly.
}
