% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance between two matrices in Rcpp}
\usage{
euc_dist(x, y)
}
\arguments{
\item{x}{NumericMatrix column should indicate variable}

\item{y}{NumericMatrix column should indicate variable}
}
\value{
double
}
\description{
This function computes a euclidean distance between two multivariate series using Rcpp.
}
