% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_nnd}
\alias{euc_nnd}
\title{Euclidean NND between validation and training blocks}
\usage{
euc_nnd(x, y)
}
\arguments{
\item{x}{NumericMatrix validation set.}

\item{y}{NumericMatrix training set.}
}
\value{
NumericVector length identical to the validation set
}
\description{
This function computes NND corresponding to euclidean distance between validation sets and training sets built by cross-validation.
}
\details{
Consider input x and y. Compute
\deqn{\sqrt{\sum (x_{ij} - y_{ij})^2}}
for each observation.
Calculate Euclidean distance between a point in validation series versus each point in training series.
Find minimum value. It is NND.
}
