% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_pdf}
\alias{euc_pdf}
\title{Euclidean pdf in Rcpp}
\usage{
euc_pdf(x, partition, display_progress = FALSE)
}
\arguments{
\item{x}{NumericMatrix, column should indicate variable.}

\item{partition}{int, equally partitioning the series.}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
NumericVector, NND pdf
}
\description{
This function computes a euclidean NND pdf of two multivariate series using Rcpp. See details what it is.
}
\details{
Implement k-fold cross-validation. Here, k is partition.
First partitioning the series equally.
One block is validation set, and the others are trainig.
Next for each validation series, it calculates sqrt(sum((x_i - x_j)^2)) versus training series.
Find the minimum result for each validation block. This is NND of each block.
Finally, you can get NND for every block and this is pdf for NND.
For \code{\link{detect_anomaly}}, this pdf is able to threshold.
Threshold is a tail of pdf, e.g. 0.99.
}
\references{
Yun, J.-H., Hwang, Y., Lee, W., Ahn, H.-K., & Kim, S.-K. (2018). \emph{Statistical Similarity of Critical Infrastructure Network Traffic Based on Nearest Neighbor Distances} (Vol. 11050, pp. 1–23). Presented at the Research in Attacks, Intrusions, and Defenses, Cham: Springer International Publishing. \url{http://doi.org/10.1007/978-3-030-00470-5_27}
}
\seealso{
\code{\link{euc_nnd}}
   \code{\link{nnd_thr}}
   \code{\link{detect_anomaly}}
}
