% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expand_label}
\alias{expand_label}
\title{Expand Label into Window}
\usage{
expand_label(x, win)
}
\arguments{
\item{x}{LogicalVector detected label}

\item{win}{int Window size}
}
\value{
LogicalVector
}
\description{
When detection label is given only one value per window,
this function expands it.

Warning: Do not use jump parameter against confusion.
}
\details{
In the window, the values are the same.
}
