% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kl_faston}
\alias{kl_faston}
\title{Simple Onlie KL Algorithm}
\usage{
kl_faston(x, newx, win, jump, lambda_p, eps, display_progress = FALSE)
}
\arguments{
\item{x}{NumericVector. univariate data set that consits of non-anomaly.}

\item{newx}{NumericVector. univariate data set that has possibility of anomaly.}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{lambda_p}{double initializing lambda_p for the threshold.}

\item{eps}{double initializing epsilon for the threshold.}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
List,
First element is kl divergence named divergence.
Second element is threshold (lambda) for detecting anomaly named threshold.
}
\description{
This function simplifies Online KL algorithm \code{\link{kl_online}}.
}
\details{
Unlike \code{\link{kl_online}}, this uses initial normal set for the first non-anomaly density estimation.
The density does not change until the algorithm ends.
}
\references{
Cho, J., Tariq, S., Lee, S., Kim, Y. G., & Woo, S. (2019). \emph{Contextual Anomaly Detection by Correlated Probability Distributions using Kullback-Leibler Divergence}. Workshop on Mining and Learning From Time Series. \url{http://doi.org/10.1145/nnnnnnn.nnnnnnn}
}
\seealso{
\link[stats]{density.default}
    \code{\link{est_density}}
    \code{\link{density_cpp}}
    \code{\link{compute_kl}}
    \code{\link{kl_dynamic}}
    \code{\link{kl_online}}
}
