% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kl_fix}
\alias{kl_fix}
\title{Fixed lambda Algorithm}
\usage{
kl_fix(x, win, jump, lambda, display_progress = FALSE)
}
\arguments{
\item{x}{NumericVector univariate data set.}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{lambda}{double threshold of K-L divergence.}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
NumericVector
}
\description{
This function implement dynamic \eqn{\lambda} algorithm for K-L divergence based on gaussianity.
}
\details{
Slide windows. In each window, estimate a density based on \link[stats]{density.default}.
Between two neighbored window, K-L divergence from current density to previous density can be computed using \code{\link{compute_kl}}.
If K-L divergence is less than \eqn{\lambda}, update the threshold by
}
\references{
Cho, J., Tariq, S., Lee, S., Kim, Y. G., & Woo, S. (2019). \emph{Contextual Anomaly Detection by Correlated Probability Distributions using Kullback-Leibler Divergence}. Workshop on Mining and Learning From Time Series. \url{http://doi.org/10.1145/nnnnnnn.nnnnnnn}
}
\seealso{
\link[stats]{density.default}
    \code{\link{est_density}}
    \code{\link{density_cpp}}
    \code{\link{compute_kl}}
}
