% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_kl}
\alias{match_kl}
\title{Matching KL divergence label to individual observation}
\usage{
match_kl(d, win)
}
\arguments{
\item{d}{LogicalVector anomaly of \code{\link{kl_dynamic}} or the result of detection by fixed algorithm}

\item{win}{int window size.}
}
\value{
NumericVector of number identical to the original series except the last window.
}
\description{
Give KL divergence anomaly prediction of each window to individual observation.
}
\details{
This function is not appropriate when jump option is used.
In other words, use only when the series has been partitioned.
}
\seealso{
\code{\link{kl_fix}}
    \code{\link{kl_dynamic}}
}
