% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nns_cpp}
\alias{nns_cpp}
\title{Windowed NNS in Rcpp}
\usage{
nns_cpp(data, win)
}
\arguments{
\item{data}{NumericMatrix multivariate data set}

\item{win}{int window size for sliding window}
}
\value{
NumericMatrix
}
\description{
This function computes a windowed NNS.
Compute NND sliding window across given series.
}
\details{
Given n x p data, slide a window.
Compute NND for each pair of moving window.
}
