% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_nnd}
\alias{pred_nnd}
\title{Compute NND for Online Data-set}
\usage{
pred_nnd(data, newdata, win, jump, d, display_progress = FALSE)
}
\arguments{
\item{data}{NumericMatrix. original data set.}

\item{newdata}{NumericMatrix. updated data set.}

\item{win}{int. window size.}

\item{jump}{int. shift size.}

\item{d}{Function. distance function. By default, \code{\link{compute_euc}}. Arguments must be two matrices and the output double.}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
NumericVector. NND for each window.
}
\description{
This function computes NND of the data set which consists of anomaly.
}
\details{
In the window in new data set, scan the original set sliding window.
Minimum distance is the NND of that window.
Repeat for every window.
}
