% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_som.R
\name{reduce_mds}
\alias{reduce_mds}
\title{Matrix Representation of Sliding Windows using MDS}
\usage{
reduce_mds(x, win, jump, parallel = FALSE)
}
\arguments{
\item{x}{mat multivariate series.}

\item{win}{int window size.}

\item{jump}{int jump size of sliding window.}

\item{parallel}{If TRUE, use parallel operation with \link[foreach]{foreach}.}
}
\value{
mat
}
\description{
Bind every window by row,
using multidimensional scaling in each window.
}
\details{
Each column represents an observation in each window.
Each row represents an distinct window.
Given the size of window win and the jump size jump, the number of the row becomes (nrow - win) / jump + 1.
The data set is multivariate series, in general, so we perform dimension reduction using MDS.
}
