% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduce_pca}
\alias{reduce_pca}
\title{Matrix Representation of Sliding Windows using PCA}
\usage{
reduce_pca(x, win, jump)
}
\arguments{
\item{x}{mat multivariate series.}

\item{win}{int window size.}

\item{jump}{int jump size of sliding window.}
}
\value{
mat
}
\description{
Bind every window by row,
using principal components analysis in each window.
}
\details{
Each column represents an observation in each window.
Each row represents an distinct window.
Given the size of window win and the jump size jump, the number of the row becomes (nrow - win) / jump + 1.
The data set is multivariate series, in general, so we perform dimension reduction using PCA.
}
