% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_erase}
\alias{row_erase}
\title{Remove row index of a matrix in C++}
\usage{
row_erase(x, rowID)
}
\arguments{
\item{x}{NumericMatrix}

\item{rowID}{IntegerVector row ids to be removed.}
}
\value{
NumericMatrix
}
\description{
This function removes a row index of NumericMatrix in Rcpp.
}
\references{
\url{https://stackoverflow.com/questions/33507695/rcpp-numericmatrix-how-to-erase-a-row-column}
}
