% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_rcpp}
\alias{seq_rcpp}
\title{Sequence by 1 in Rcpp}
\usage{
seq_rcpp(from, to)
}
\arguments{
\item{from}{int the starting value of the sequence.}

\item{to}{int the end value of the sequence.}
}
\value{
IntegerVector
}
\description{
This function generates a integer sequence with increment of 1 in Rcpp.
}
