% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatanomaly-package.R
\docType{package}
\name{swatanomaly-package}
\alias{swatanomaly-package}
\alias{_PACKAGE}
\alias{swatanomaly}
\title{swatanomaly: various algorithms for anomaly detection}
\description{
swatanomaly provides some algorithms used by NSR anomaly detection projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim}
}

}
\author{
Young-geun Kim <\url{https://github.com/ygeunkim}>
}
\keyword{internal}
