% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_static.R
\name{train_cusum}
\alias{train_cusum}
\title{Choose Threshold for CUSUM}
\usage{
train_cusum(x, win, jump, norm, prob = 0.9, display_progress = FALSE)
}
\arguments{
\item{x}{Train data}

\item{win}{window size}

\item{jump}{jump size}

\item{norm}{p-norm}

\item{prob}{Quantile}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
Suggested threshold value for \code{\link{detect_cusum}}
}
\description{
Choose Threshold for CUSUM
}
\details{
By computing the quantile of every CUSUM, threshold is given.
}
\seealso{
\code{\link{compute_cusum}}
\code{\link{detect_cusum}}
}
