% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{win_mat}
\alias{win_mat}
\title{Matrix Representation of Sliding Windows}
\usage{
win_mat(x, win, jump)
}
\arguments{
\item{x}{NumericVector univariate data.}

\item{win}{int window size.}

\item{jump}{int jump size of sliding window.}
}
\value{
NumericMatrix
}
\description{
Bind every window by row.
}
\details{
Each column represents an observation in each window.
Each row represents an distinct window.
Given the size of window win and the jump size jump, the number of the row becomes (nrow - win) / jump + 1.
}
