% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_nnd}
\alias{detect_nnd}
\title{Anomaly detection using NND}
\usage{
detect_nnd(data, win, thr)
}
\arguments{
\item{data}{NumericMatrix multivariate data set}

\item{win}{int window size for sliding window}

\item{thr}{threshold for anomaly detection, in each window}
}
\value{
LogicalVector,
If NND is (strictly) larger than threshold then TRUE.
Otherwise, FALSE
}
\description{
This function detects anomaly based on NND.
}
\details{
Given n x p data, slide a window.
Compute NND for each pair of moving window.
For threshold, users can use tail value of \code{\link{euc_pdf}}.
}
