% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detect_nndvec}
\alias{detect_nndvec}
\title{Anomaly detection after conducting NND}
\usage{
detect_nndvec(nnd, win, thr)
}
\arguments{
\item{nnd}{NumericVector result of \code{\link{nns_cpp}}}

\item{win}{int window size for sliding window}

\item{thr}{threshold for anomaly detection, in each window}
}
\value{
LogicalVector,
If NND is (strictly) larger than threshold then TRUE.
Otherwise, FALSE
}
\description{
This function detects anomaly based on NND, given \code{\link{nns_cpp}}.
}
\details{
Given n x p data, slide a window.
Compute NND for each pair of moving window.
For threshold, users can use tail value of \code{\link{euc_pdf}}.
}
