% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_dist}
\alias{euc_dist}
\title{Euclidean distance between two matrices in Rcpp}
\usage{
euc_dist(x, y)
}
\arguments{
\item{x}{NumericMatrix column should indicate variable}

\item{y}{NumericMatrix column should indicate variable}
}
\value{
double
}
\description{
This function computes a euclidean distance between two multivariate series using Rcpp.
}
\details{
For input x and y, compute
\deqn{\sum \sqrt{\sum (x_{ij} - y_{ij})^2}}
At first, the function calculates Euclidean distance pairwisely.
After that, sum over every observation.
}
