% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euc_pdf}
\alias{euc_pdf}
\title{Euclidean pdf in Rcpp}
\usage{
euc_pdf(x, y, partition)
}
\arguments{
\item{x}{NumericMatrix column should indicate variable}

\item{y}{NumericMatrix column should indicate variable}

\item{partition}{int partitioning equally the series}
}
\value{
NumericVector NND for each block
}
\description{
This function computes a euclidean NND pdf of two multivariate series using Rcpp. See details what it is.
}
\details{
First partitioning the series equally.
Next for each partitioned block, it calculates sqrt(sum((x_i - y_i)^2)) versus the other blocks.
Find the minimum result for each block. This is NND of each block.
}
