% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_kl}
\alias{match_kl}
\title{Matching KL divergence to individual observation}
\usage{
match_kl(d, win, jump, last_win = FALSE)
}
\arguments{
\item{d}{NumeriVector kl divergence vector.}

\item{win}{int window size.}

\item{jump}{int jump size for sliding window.}

\item{last_win}{Fill last window? If TRUE, fill the last window with same value with the KL of the former window. Otherwise, leave them. By default, FALSE.}
}
\value{
NumericVector of number identical to the original series except the last window.
}
\description{
Give KL divergence values of each window to individual observation.
}
\seealso{
\code{\link{kl_fix}}
    \code{\link{kl_dynamic}}
}
