% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_nnd.R
\name{nnd_thr}
\alias{nnd_thr}
\title{NND pdf Threshold}
\usage{
nnd_thr(data, part_size, prob, display_progress = FALSE)
}
\arguments{
\item{data}{NumericMatrix multivariate data set}

\item{part_size}{partition for \code{\link{euc_pdf}}}

\item{prob}{probability for quantile. See \link[stats]{quantile.default}.}

\item{display_progress}{If TRUE, display a progress bar. By default, FALSE.}
}
\value{
double quantile value for the result of \code{\link{euc_pdf}}
}
\description{
This function computes threshold using \code{\link{euc_pdf}}
}
\details{
Sometimes you might use NND pdf for threshold.
Threshold is a tail of pdf, e.g. 0.99.
}
\seealso{
\code{\link{euc_pdf}}
   \link[stats]{quantile.default}
   \code{\link{detect_nndvec}}
}
