% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{partnnd}
\alias{partnnd}
\title{Sliding window for NND}
\usage{
partnnd(data, win, jump, partition, base_id, d)
}
\arguments{
\item{data}{NumericMatrix. data to be calculated NND.}

\item{win}{int. window size.}

\item{jump}{int. shift size.}

\item{partition}{NumericVector. indices that can indicate partitions.}

\item{base_id}{int. An index of chosen partition among partiton vector.}

\item{d}{Function. distance function.}
}
\value{
NumericVector. NND vector for each window in the chosen partition.
}
\description{
For the chosen partition, get NND for every pair of window.
This function is constructed for the other function.
}
\seealso{
\code{\link{nnd_normal}}
    \code{\link{pred_nnd}}
}
