% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_static.R
\name{train_norm}
\alias{train_norm}
\title{Choose Threshold for p-norm}
\usage{
train_norm(x, norm, prob = 0.9)
}
\arguments{
\item{x}{Train data}

\item{norm}{p-norm}

\item{prob}{Quantile}
}
\value{
Suggested threshold value for \code{\link{detect_norm}}
}
\description{
Choose Threshold for p-norm
}
\details{
By computing the quantile of every p-norm, threshold is given.
}
\seealso{
\code{\link{compute_norm}}
\code{\link{detect_norm}}
}
